fix static partition sorting
