+++ python_modules/dagster/dagster/_core/definitions/partition.py
@@ -438,8 +438,6 @@ class StaticPartitionsDefinition(PartitionsDefinition[str]):
         partition_keys = self.get_partition_keys(
             current_time=current_time, dynamic_partitions_store=dynamic_partitions_store
         )
-        if not ascending:
-            partition_keys = list(reversed(partition_keys))
         return Connection.create_from_sequence(
             partition_keys, limit=limit, ascending=ascending, cursor=cursor
         )
